ShoppingCart functionality

+++ FrietZaak.Server/Models/Order.cs
@@ -6,7 +6,7 @@
         public int? CustomerId { get; set; } // nullable want kunt bestellen zonder customer account
         public Customer? Customer { get; set; }
         public List<OrderLine>? Items { get; set; }
-        public float TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public float Discount { get; set; }
         public bool Finished { get; set; }
